Refactored the monome display logic.
Whereas the refactor before this focused on how rows are flagged to be
processed on quantize, this refactor push reworked how they are
displayed.

rove.c:monome_display_loop() now uses an approach similar to
rove_jack.c:process(), first iterating through the groups and displaying
their active files, then stepping through a bitmask in the rove_monome_t
structure and displaying any files that have expressly requested it.

This also allows for getting rid of the FILE_STATUS_DEACTIVATE constant,
which is something that I've wanted to do for a while.  The rove_file_t
is now effectively a state machine with two states, FILE_STATUS_ACTIVE
and FILE_STATUS_INACTIVE, and the state transitions are now handled in a
much more robust fashion.